+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/DefaultCardEditUIHandler.kt
@@ -95,6 +95,5 @@ internal class DefaultCardEditUIHandler(
                 onCardValuesChanged = onCardValuesChanged
             )
         }
-
     }
 }
