+++ paymentsheet/src/test/java/com/stripe/android/model/PaymentMethodFixtures.kt
@@ -15,7 +15,7 @@ import java.util.UUID
 import java.util.concurrent.ThreadLocalRandom
 
 internal object PaymentMethodFixtures {
-    private val CARD = PaymentMethod.Card(
+    val CARD = PaymentMethod.Card(
         brand = CardBrand.Visa,
         checks = PaymentMethod.Card.Checks(
             addressLine1Check = "unchecked",
