+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/FakeCardEditUIHandlerFactory.kt
@@ -23,5 +23,4 @@ internal class FakeCardEditUIHandlerFactory : CardEditUIHandler.Factory {
             this.cardEditUIHandler = it
         }
     }
-
 }
